requirements.txt
pillow
click

